:recycle: Refactor poject selector
