+++ classifai-core/src/main/java/ai/classifai/selector/project/ProjectImportSelector.java
@@ -17,14 +17,12 @@ package ai.classifai.selector.project;
 
 import ai.classifai.action.ActionConfig;
 import ai.classifai.action.ProjectImport;
-import ai.classifai.ui.launcher.LogoLauncher;
+import ai.classifai.ui.SelectionWindow;
 import ai.classifai.ui.launcher.WelcomeLauncher;
-import ai.classifai.util.ParamConfig;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FilenameUtils;
 
 import javax.swing.*;
-import javax.swing.filechooser.FileNameExtensionFilter;
 import java.awt.*;
 import java.io.File;
 import java.nio.file.Paths;
@@ -37,18 +35,7 @@ import static javax.swing.JOptionPane.showMessageDialog;
  * @author codenamewei
  */
 @Slf4j
-public class ProjectImportSelector
-{
-    private enum ImportSelectionWindowStatus
-    {
-        WINDOW_OPEN,
-        WINDOW_CLOSE
-    }
-    private ImportSelectionWindowStatus windowStatus = ImportSelectionWindowStatus.WINDOW_CLOSE;
-
-    private static final FileNameExtensionFilter imgfilter = new FileNameExtensionFilter(
-            "Json Files", "json");
-
+public class ProjectImportSelector extends SelectionWindow {
     public void run()
     {
         try
@@ -61,8 +48,8 @@ public class ProjectImportSelector
                     {
                         windowStatus = ImportSelectionWindowStatus.WINDOW_OPEN;
 
-                        JFrame frame = initiateFrame();
-                        JFileChooser chooser = initiateChooser();
+                        JFrame frame = initFrame();
+                        JFileChooser chooser = initChooser(JFileChooser.FILES_ONLY);
 
                         //Important: prevent Welcome Console from popping out
                         WelcomeLauncher.setToBackground();
@@ -104,44 +91,6 @@ public class ProjectImportSelector
         {
             log.info("ProjectHandler for File type failed to open", e);
         }
-    }
-
-    private JFrame initiateFrame()
-    {
-        Point pt = MouseInfo.getPointerInfo().getLocation();
-        JFrame frame = new JFrame();
-        frame.setIconImage(LogoLauncher.getClassifaiIcon());
-
-        frame.setAlwaysOnTop(true);
-        frame.setDefaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE);
-        frame.setLocation(pt);
-        frame.requestFocus();
-        frame.setVisible(false);
-
-        return frame;
-    }
-
-    private JFileChooser initiateChooser()
-    {
-        JFileChooser chooser = new JFileChooser() {
-            @Override
-            protected JDialog createDialog(Component parent)
-                    throws HeadlessException {
-                JDialog dialog = super.createDialog(parent);
-                dialog.setLocationByPlatform(true);
-                dialog.setAlwaysOnTop(true);
-                return dialog;
-            }
-        };
-
-        chooser.setCurrentDirectory(ParamConfig.getRootSearchPath());
-        chooser.setFileFilter(imgfilter);
-        chooser.setDialogTitle("Select Files");
-        chooser.setMultiSelectionEnabled(false);
-        chooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
-        chooser.setAcceptAllFileFilterUsed(false);
-
-        return chooser;
     }
 
     private void showAbortImportPopup()
